+++ lib/rsgem/errors/base.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module RSGem
+  module Errors
+    class Base < StandardError
+    end
+  end
+end