+++ lib/rsgem/errors/missing_gem_name_error.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module RSGem
-  class MissingGemNameError < StandardError
-  end
-end