+++ lib/rsgem/tasks/run_rubocop.rb
@@ -3,16 +3,13 @@
 module RSGem
   module Tasks
     class RunRubocop < Base
+      OUTPUT = OutputStruct.new(name: 'Run rubocop')
+
       def perform
-        puts "\tRubocop:"
-        @output = `cd #{context.folder_path} && bundle exec rubocop -A`
-        puts "\t\t#{last_line}"
-      end
+        return if system("cd #{context.folder_path} && bundle exec rubocop -A",
+                         %i[out err] => '/dev/null')
 
-      private
-
-      def last_line
-        @output.split("\n").last
+        raise RSGem::Errors::Base, 'Failed to run `bundle exec rubocop -A\''
       end
     end
   end