+++ spec/gem_spec.rb
@@ -105,16 +105,6 @@ RSpec.describe RSGem::Gem do
       it 'adds license file with Rootstrap name' do
         expect(license).to include "Copyright (c) #{Date.today.year} Rootstrap"
       end
-
-      context 'running inside the new gem' do
-        before { Dir.chdir(gem_name) }
-        after { Dir.chdir('../') }
-
-        it 'makes the code analysis task pass' do
-          load File.expand_path("../#{gem_name}/Rakefile", __dir__)
-          expect { Rake::Task['code_analysis'].invoke }.not_to raise_error
-        end
-      end
     end
 
     context 'with github actions as ci provider' do
@@ -139,7 +129,7 @@ RSpec.describe RSGem::Gem do
 
     context 'when gem name options is missing' do
       it 'raises a missing gem name error' do
-        expect { described_class.new({}).create }.to raise_error(RSGem::MissingGemNameError)
+        expect { described_class.new({}).create }.to raise_error(RSGem::Errors::MissingGemName)
       end
     end
 
@@ -162,7 +152,9 @@ RSpec.describe RSGem::Gem do
 
       it 'exits the process' do
         expect_any_instance_of(Object).to receive(:system).and_return(false)
-        expect { described_class.new(gem_name: gem_name).create }.to raise_error(SystemExit)
+        expect { described_class.new(gem_name: gem_name).create }.to(
+          raise_error(RSGem::Errors::Base)
+        )
       end
     end
   end