+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # master
 
+* Using Github Actions as the default CI provider
 * Dropping Ruby 2.4 support
 * New CLI output format
-
-    *Juan Manuel Ramallo*
-
 * Add rubocop-rootstrap
 
     *Juan Manuel Ramallo*